Create separate Dirs for Classes and Enums.
