+++ ChessProject-Csharp/src/Enums/MovementType.cs
@@ -1,0 +1,8 @@
+﻿namespace SolarWinds.MSP.Chess
+{
+    public enum MovementType
+    {
+        Move,
+        Capture
+    }
+}
