+++ ChessProject-Csharp/src/Enums/PieceColor.cs
@@ -1,0 +1,8 @@
+﻿namespace SolarWinds.MSP.Chess
+{
+    public enum PieceColor
+    {
+        Black,
+        White
+    }
+}
