--- ChessProject-Csharp/src/MovementType.cs
@@ -1,8 +1,0 @@
-﻿namespace SolarWinds.MSP.Chess
-{
-    public enum MovementType
-    {
-        Move = 0,
-        Capture = 1
-    }
-}
