--- ChessProject-Csharp/src/PieceColor.cs
@@ -1,8 +1,0 @@
-﻿namespace SolarWinds.MSP.Chess
-{
-    public enum PieceColor
-    {
-        Black = 0,
-        White = 1
-    }
-}
